Add styling to the back link
